+++ styled_text/lib/styled_text.dart
@@ -6,8 +6,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:xmlstream/xmlstream.dart';
 
-typedef ActionTappedCallback = void Function(
-    TextSpan text, Map<String, String> attributes);
+typedef ActionTappedCallback = void Function(TextSpan text, Map<String, String> attributes);
 
 class _StyledTextRecoginzer extends TapGestureRecognizer {
   ActionTappedCallback onTextTap;
@@ -86,9 +85,9 @@ class IconStyle extends TextStyle {
 
 ///
 /// Text widget with formatting via tags.
-/// 
+///
 /// Formatting is specified as xml tags. For each tag, you can specify a style in the [styles] parameter.
-/// 
+///
 /// Example:
 /// ```dart
 /// StyledText(
@@ -99,26 +98,37 @@ class IconStyle extends TextStyle {
 /// )
 /// ```
 /// See also:
-/// 
+///
 /// * [TextStyle], which discusses how to style text.
-/// 
+///
 class StyledText extends StatefulWidget {
-
   /// The text to display in this widget. The text must be valid xml.
+  ///
+  /// You need to escape specific XML characters in text:
+  ///
+  /// ```
+  /// Original character  Escaped character
+  /// ------------------  -----------------
+  /// "                   &quot;
+  /// '                   &apos;
+  /// &                   &amp;
+  /// <                   &lt;
+  /// >                   &gt;
+  /// ```
+  ///
   final String text;
-
-  /// Do not ignore line feeds in the source text.
-  @Deprecated('will be removed in 1.0.2, use newLineAsBreaks instead')
-  final bool isNewLineAsBreaks;
 
   /// Treat newlines as line breaks.
   final bool newLineAsBreaks;
+
+  /// Is text selectable?
+  final bool selectable;
 
   /// Default text style.
   final TextStyle style;
 
   /// Style map for tags in text.
-  /// 
+  ///
   /// Example:
   /// ```dart
   /// StyledText(
@@ -170,12 +180,11 @@ class StyledText extends StatefulWidget {
   /// {@macro flutter.painting.textPainter.strutStyle}
   final StrutStyle strutStyle;
 
-  /// Create a text widget with formatting via tags. 
+  /// Create a text widget with formatting via tags.
   ///
   StyledText({
     Key key,
     @required this.text,
-    @Deprecated('will be removed in 1.0.2, use newLineAsBreaks instead') this.isNewLineAsBreaks = false,
     this.newLineAsBreaks = false,
     this.style,
     @required this.styles,
@@ -187,7 +196,85 @@ class StyledText extends StatefulWidget {
     this.maxLines,
     this.locale,
     this.strutStyle,
-  }) : super(key: key);
+  })  : this.selectable = false,
+        this._focusNode = null,
+        this._showCursor = false,
+        this._autofocus = false,
+        this._toolbarOptions = null,
+        this._cursorWidth = null,
+        this._cursorHeight = null,
+        this._cursorRadius = null,
+        this._cursorColor = null,
+        this._dragStartBehavior = DragStartBehavior.start,
+        this._enableInteractiveSelection = false,
+        this._onTap = null,
+        this._scrollPhysics = null,
+        this._textHeightBehavior = null,
+        this._textWidthBasis = null,
+        super(key: key);
+
+  /// Create a selectable text widget with formatting via tags.
+  ///
+  /// See [SelectableText.rich] for more options.
+  StyledText.selectable(
+      {Key key,
+      @required this.text,
+      this.newLineAsBreaks = false,
+      this.style,
+      @required this.styles,
+      this.textAlign = TextAlign.start,
+      this.textDirection,
+      this.textScaleFactor,
+      this.maxLines,
+      this.strutStyle,
+      FocusNode focusNode,
+      bool showCursor = false,
+      bool autofocus = false,
+      ToolbarOptions toolbarOptions,
+      double cursorWidth = 2.0,
+      double cursorHeight,
+      Radius cursorRadius,
+      Color cursorColor,
+      DragStartBehavior dragStartBehavior = DragStartBehavior.start,
+      bool enableInteractiveSelection = true,
+      GestureTapCallback onTap,
+      ScrollPhysics scrollPhysics,
+      TextHeightBehavior textHeightBehavior,
+      TextWidthBasis textWidthBasis})
+      : this.selectable = true,
+        this.softWrap = true,
+        this.overflow = TextOverflow.clip,
+        this.locale = null,
+        this._focusNode = focusNode,
+        this._showCursor = showCursor,
+        this._autofocus = autofocus,
+        this._toolbarOptions = toolbarOptions,
+        this._cursorWidth = cursorWidth,
+        this._cursorHeight = cursorHeight,
+        this._cursorRadius = cursorRadius,
+        this._cursorColor = cursorColor,
+        this._dragStartBehavior = dragStartBehavior,
+        this._enableInteractiveSelection = enableInteractiveSelection,
+        this._onTap = onTap,
+        this._scrollPhysics = scrollPhysics,
+        this._textHeightBehavior = textHeightBehavior,
+        this._textWidthBasis = textWidthBasis,
+        super(key: key);
+
+  final FocusNode _focusNode;
+  final bool _showCursor;
+  final bool _autofocus;
+  final ToolbarOptions _toolbarOptions;
+  final double _cursorWidth;
+  final double _cursorHeight;
+  final Radius _cursorRadius;
+  final Color _cursorColor;
+  final DragStartBehavior _dragStartBehavior;
+  final bool _enableInteractiveSelection;
+  final GestureTapCallback _onTap;
+  final ScrollPhysics _scrollPhysics;
+  final TextHeightBehavior _textHeightBehavior;
+  final TextWidthBasis _textWidthBasis;
 
   @override
   _StyledTextState createState() => _StyledTextState();
@@ -210,8 +297,7 @@ class _StyledTextState extends State<StyledText> {
     if ((widget.text != oldWidget.text) ||
         (widget.styles != oldWidget.styles) ||
         (widget.style != oldWidget.style) ||
-        (widget.newLineAsBreaks != oldWidget.newLineAsBreaks) ||
-        (widget.isNewLineAsBreaks != oldWidget.isNewLineAsBreaks)) {
+        (widget.newLineAsBreaks != oldWidget.newLineAsBreaks)) {
       _updateTextSpans(force: true);
     }
   }
@@ -223,7 +309,7 @@ class _StyledTextState extends State<StyledText> {
 
       String textValue = _text;
 
-      if (widget.newLineAsBreaks || widget.isNewLineAsBreaks) {
+      if (widget.newLineAsBreaks) {
         textValue = textValue.replaceAll("\n", '<br/>');
       }
 
@@ -234,8 +320,7 @@ class _StyledTextState extends State<StyledText> {
       ListQueue<TextSpan> textQueue = ListQueue();
       Map<String, String> attributes;
 
-      var xmlStreamer = new XmlStreamer(
-          '<?xml version="1.0" encoding="UTF-8"?><root>' + textValue + '</root>');
+      var xmlStreamer = new XmlStreamer('<?xml version="1.0" encoding="UTF-8"?><root>' + textValue + '</root>');
       xmlStreamer.read().listen((e) {
         switch (e.state) {
           case XmlState.Text:
@@ -268,13 +353,11 @@ class _StyledTextState extends State<StyledText> {
                   ),
                 );
               } else {
-                final _StyledTextRecoginzer recognizer =
-                    ((style is ActionTextStyle) && style.onTap != null)
-                        ? _StyledTextRecoginzer(onTextTap: style.onTap)
-                        : null;
+                final _StyledTextRecoginzer recognizer = ((style is ActionTextStyle) && style.onTap != null)
+                    ? _StyledTextRecoginzer(onTextTap: style.onTap)
+                    : null;
 
-                node = TextSpan(
-                    style: style, children: [], recognizer: recognizer);
+                node = TextSpan(style: style, children: [], recognizer: recognizer);
               }
             }
 
@@ -318,19 +401,46 @@ class _StyledTextState extends State<StyledText> {
 
   @override
   Widget build(BuildContext context) {
-    return (_textSpans != null)
-        ? RichText(
-            textAlign: widget.textAlign,
-            textDirection: widget.textDirection,
-            softWrap: widget.softWrap,
-            overflow: widget.overflow,
-            textScaleFactor: widget.textScaleFactor ??
-                MediaQuery.of(context).textScaleFactor,
-            maxLines: widget.maxLines,
-            locale: widget.locale,
-            strutStyle: widget.strutStyle,
-            text: _textSpans,
-          )
-        : const SizedBox();
+    if (_textSpans == null) return const SizedBox();
+
+    if (!widget.selectable) {
+      return RichText(
+        textAlign: widget.textAlign,
+        textDirection: widget.textDirection,
+        softWrap: widget.softWrap,
+        overflow: widget.overflow,
+        textScaleFactor: widget.textScaleFactor ?? MediaQuery.of(context).textScaleFactor,
+        maxLines: widget.maxLines,
+        locale: widget.locale,
+        strutStyle: widget.strutStyle,
+        text: _textSpans,
+      );
+    } else {
+      return SelectableText.rich(
+        _textSpans,
+        focusNode: widget._focusNode,
+        showCursor: widget._showCursor,
+        autofocus: widget._autofocus,
+        toolbarOptions: widget._toolbarOptions,
+        cursorWidth: widget._cursorWidth,
+        cursorHeight: widget._cursorHeight,
+        cursorRadius: widget._cursorRadius,
+        cursorColor: widget._cursorColor,
+        dragStartBehavior: widget._dragStartBehavior,
+        enableInteractiveSelection: widget._enableInteractiveSelection,
+        onTap: widget._onTap,
+        scrollPhysics: widget._scrollPhysics,
+        textHeightBehavior: widget._textHeightBehavior,
+        textWidthBasis: widget._textWidthBasis,
+        textAlign: widget.textAlign,
+        textDirection: widget.textDirection,
+        // softWrap
+        // overflow
+        textScaleFactor: widget.textScaleFactor ?? MediaQuery.of(context).textScaleFactor,
+        maxLines: widget.maxLines,
+        // locale
+        strutStyle: widget.strutStyle,
+      );
+    }
   }
 }